Ported all unit tests to Google Test. Got rid of minunit.
