+++ test/ClusterChainFileTest.cpp
@@ -1,6 +1,5 @@
 #include <errno.h>
-
-#include "minunit.h"
+#include "gtest/gtest.h"
 #include "../include/common.h"
 #include "../include/alist.h"
 #include "../include/cch.h"
@@ -8,34 +7,36 @@
 #include "../include/cchfile.h"
 #include "../include/lfnde.h"
 
-static const char *G_DISK_FNAME = "/home/pavel/projects/vfat/test/disk0";
-
-MU_TEST_SETUP(setup)
+class ClusterChainFileTest : public ::testing::Test
 {
-    struct fdisk disk;
+protected:
+    const char *DISK_FNAME = "disk0";
 
-    fdisk_create(G_DISK_FNAME, &disk);
-    cchdir_formatdev(&disk, 1024 * 1024, 512, 1);
+    void SetUp() override
+    {
+        struct fdisk disk;
 
-    fdisk_close(&disk);
-}
+        fdisk_create(DISK_FNAME, &disk);
+        cchdir_formatdev(&disk, 1024 * 1024, 512, 1);
 
-MU_TEST_TEARDOWN(teardown)
+        fdisk_close(&disk);
+    }
+
+    void TearDown() override
+    {
+        remove(DISK_FNAME);
+        ::__vfat_errno = 0;
+    }
+};
+
+TEST_F(ClusterChainFileTest, SetLength)
 {
-    remove(G_DISK_FNAME);
-    ::__vfat_errno = 0;
-}
-
-MU_TEST(test_set_length)
-{
-    MU_PRINT_TEST_INFO();
-
     struct fdisk disk;
     struct vbr br;
     struct fat fat;
     struct cchdir root;
 
-    fdisk_open(G_DISK_FNAME, &disk);
+    fdisk_open(DISK_FNAME, &disk);
     vbr_read(&disk, &br);
     fat_read(&disk, &br, &fat);
     cchdir_readroot(&disk, &fat, &root);
@@ -46,26 +47,24 @@ MU_TEST(test_set_length)
     cchdir_addfile(&root, "index.htm", &e);
     cchdir_getfile(&root, &e, &file);
 
-    MU_ASSERT_U32_EQ(0, cchfile_getlen(&file));
+    EXPECT_EQ(0, cchfile_getlen(&file));
     cchfile_setlen(&file, 100);
-    MU_ASSERT_U32_EQ(100, cchfile_getlen(&file));
-    MU_ASSERT(cch_getsize(file.chain) >= 100);
+    EXPECT_EQ(100, cchfile_getlen(&file));
+    EXPECT_GE(cch_getsize(file.chain), 100);
 
     cchfile_destruct(&file);
     cchdir_destruct(&root);
     fdisk_close(&disk);
 }
 
-MU_TEST(test_read_write)
+TEST_F(ClusterChainFileTest, ReadWrite)
 {
-    MU_PRINT_TEST_INFO();
-
     struct fdisk disk;
     struct vbr br;
     struct fat fat;
     struct cchdir root;
 
-    fdisk_open(G_DISK_FNAME, &disk);
+    fdisk_open(DISK_FNAME, &disk);
     vbr_read(&disk, &br);
     fat_read(&disk, &br, &fat);
     cchdir_readroot(&disk, &fat, &root);
@@ -90,28 +89,17 @@ MU_TEST(test_read_write)
 
     // Read from device
     cchfile_read(&disk, &file, 0, len, &nread, readbuf);
-    MU_ASSERT_U32_EQ(len, nread);
+    EXPECT_EQ(len, nread);
 
     for (i = 0; i < len; ++i) {
-        MU_ASSERT_U32_EQ(i % 256, readbuf[i]);
+        EXPECT_EQ(i % 256, readbuf[i]);
     }
 
     // Read too long
     cchfile_read(&disk, &file, 0, len + 1, &nread, readbuf);
-    MU_ASSERT_U32_EQ(len, nread);
+    EXPECT_EQ(len, nread);
 
     cchfile_destruct(&file);
     cchdir_destruct(&root);
     fdisk_close(&disk);
-}    
-
-MU_TEST_SUITE(cchfile_test_suite)
-{
-    MU_SUITE_CONFIGURE(&setup, &teardown);
-    MU_RESET();
-
-    MU_RUN_TEST(test_set_length);
-    MU_RUN_TEST(test_read_write);
-
-    MU_REPORT();
 }
