--- test/testsuite.h
@@ -1,13 +1,0 @@
-#ifndef VFAT_TEST_SUITE_H
-#define VFAT_TEST_SUITE_H
-
-#include "minunit.h"
-
-MU_TEST_SUITE(vbr_test_suite);
-MU_TEST_SUITE(fat_test_suite);
-MU_TEST_SUITE(cch_test_suite);
-MU_TEST_SUITE(cchdir_test_suite);
-MU_TEST_SUITE(cchfile_test_suite);
-MU_TEST_SUITE(filesys_test_suite);
-
-#endif /* VFAT_TEST_SUITE_H */
